LU-12898 utils: %llu mismatch with type __u64 on ppcle64
Fix build errors like this one on ppcle64:

BUILDSTDERR: libmount_utils_zfs.c: In function 'zfs_mkfs_opts':
BUILDSTDERR: libmount_utils_zfs.c:573:5: error: format '%llu' expects
argument of type 'long long unsigned int', but argument 4 has type
'__u64' [-Werror=format=]
BUILDSTDERR:      mop->mo_device_kb * 1024);

__u64 was treated as an unsigned long long which breaks the build on
ppc64le, where they are not the same size.

In printf cases, cast to unsigned long long to match the printf format
so the format is compatible with the type and it is guaranteed
not to lose any data.

In the case of sscanf(), replace the call with strtoull() to eliminate
the issue.

Lustre-change: https://review.whamcloud.com/36558
Lustre-commit: 56b4b112a497661de8dbf5a851c7a045d470deff

Test-Parameters: trivial
Reviewed-on: https://review.whamcloud.com/37130
Tested-by: jenkins <devops@whamcloud.com>
Tested-by: Maloo <maloo@whamcloud.com>

+++ lustre/utils/lustre_lfsck.c
@@ -629,7 +629,8 @@ bad_type:
 			printf("%s_osts_%s: %d\n", name,
 			       lfsck_status2name(j), query.lu_osts_count[i][j]);
 
-		printf("%s_repaired: %llu\n", name, query.lu_repaired[i]);
+		printf("%s_repaired: %llu\n", name,
+		       (unsigned long long)query.lu_repaired[i]);
 	}
 
 	return 0;
